refactor(Gcodefiles): refactor gcodefiles panel and table
