+++ src/components/dialogs/GcodefilesCreateDirectoryDialog.vue
@@ -1,0 +1,88 @@
+<template>
+    <v-dialog :value="showDialog" width="400">
+        <panel :title="$t('Files.NewDirectory')" card-class="gcodefiles-new-directory-dialog" :margin-bottom="false">
+            <template #buttons>
+                <v-btn icon tile @click="closePrompt">
+                    <v-icon>{{ mdiCloseThick }}</v-icon>
+                </v-btn>
+            </template>
+            <v-card-text>
+                <v-text-field
+                    ref="inputFieldCreateDirectory"
+                    v-model="name"
+                    :label="$t('Files.Name')"
+                    required
+                    :rules="nameInputRules"
+                    @update:error="updateIsInvalidName"
+                    @keydown.enter="createDirectoryAction" />
+            </v-card-text>
+            <v-card-actions>
+                <v-spacer />
+                <v-btn text @click="closePrompt">{{ $t('Files.Cancel') }}</v-btn>
+                <v-btn
+                    :disabled="isInvalidName || name.length === 0"
+                    color="primary"
+                    text
+                    @click="createDirectoryAction">
+                    {{ $t('Files.Create') }}
+                </v-btn>
+            </v-card-actions>
+        </panel>
+    </v-dialog>
+</template>
+
+<script lang="ts">
+import { Component, Mixins, Prop, Ref, Watch } from 'vue-property-decorator'
+import BaseMixin from '@/components/mixins/base'
+import Panel from '@/components/ui/Panel.vue'
+import { mdiCloseThick } from '@mdi/js'
+import GcodefilesMixin from '@/components/mixins/gcodefiles'
+
+@Component({
+    components: { Panel },
+})
+export default class GcodefilesCreateDirectoryDialog extends Mixins(BaseMixin, GcodefilesMixin) {
+    mdiCloseThick = mdiCloseThick
+
+    name = ''
+    isInvalidName = false
+
+    @Prop({ type: Boolean, default: false }) showDialog!: boolean
+    @Ref('inputFieldCreateDirectory') readonly inputFieldCreateDirectory!: HTMLInputElement
+
+    nameInputRules = [
+        (value: string) => !!value || this.$t('Files.InvalidNameEmpty'),
+        (value: string) => !this.existsFilename(value) || this.$t('Files.InvalidNameAlreadyExists'),
+    ]
+
+    updateIsInvalidName(value: boolean) {
+        this.isInvalidName = value
+    }
+
+    createDirectoryAction() {
+        this.$socket.emit(
+            'server.files.post_directory',
+            { path: 'gcodes' + this.currentPath + '/' + this.name },
+            { action: 'files/getCreateDir' }
+        )
+
+        this.closePrompt()
+    }
+
+    closePrompt() {
+        this.$emit('close')
+    }
+
+    @Watch('showDialog')
+    onShowDialogChanged(newVal: boolean) {
+        if (!newVal) return
+
+        this.name = ''
+        this.isInvalidName = false
+
+        setTimeout(() => {
+            this.inputFieldCreateDirectory.focus()
+        }, 200)
+    }
+}
+</script>
