+++ src/store/files/types.ts
@@ -43,10 +43,8 @@ export interface FileStateFile {
 }
 
 export interface FileStateGcodefile extends FileStateFile {
+    full_filename: string
     preheat_gcode: string | null
-    small_thumbnail: string | null
-    big_thumbnail: string | null
-    big_thumbnail_width: number | null
     count_printed: number
     last_end_time: Date | null
     last_filament_used: number | null
